third_party/WebKit/LayoutTests/inspector/sources/debugger-async/async-callstack-events.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>

function testFunction()
{
    for (var name in window) {
        if (/^doTest[A-Z]/.test(name) && typeof window[name] === "function")
            window[name]();
    }
}

function setSelection(start, end)
{
    var node = document.getElementById("content").firstChild;
    var range = document.createRange();
    range.setStart(node, start);
    range.setEnd(node, end);
    var selection = window.getSelection();
    selection.removeAllRanges();
    if (start !== end)
        selection.addRange(range);
}

function doTestSelectionChange()
{
    setSelection(0, 0);
    document.addEventListener("selectionchange", onSelectionChange, false);
    setSelection(0, 4);
    setSelection(0, 8);
    setSelection(0, 0);
}

function onSelectionChange()
{
    document.removeEventListener("selectionchange", onSelectionChange, false);
    debugger;
}

function doTestHashChange()
{
    window.addEventListener("hashchange", onHashChange1, false);
    window.addEventListener("hashchange", onHashChange2, true);
    location.hash = location.hash + "x";
}

function onHashChange1()
{
    window.removeEventListener("hashchange", onHashChange1, false);
    debugger;
}

function onHashChange2()
{
    window.removeEventListener("hashchange", onHashChange2, true);
    debugger;
}

function doTestMediaEvents()
{
    var video = document.getElementById("video");
    video.addEventListener("play", onVideoPlay, false);
    video.play();
}

function onVideoPlay()
{
    video.removeEventListener("play", onVideoPlay, false);
    debugger;
}

var test = function()
{
    var totalDebuggerStatements = 4;
    InspectorTest.runAsyncCallStacksTest(totalDebuggerStatements);
}

</script>
</head>

<body onload="runTest()">
<video id="video" src="../../../media/content/test.ogv"></video>
<p id="content">
Tests asynchronous call stacks for various DOM events.
</p>
</body>
</html>
