third_party/WebKit/LayoutTests/inspector/sources/debugger-async/async-callstack-events-expected.txt

Tests asynchronous call stacks for various DOM events.

Set timer for test function.
Captured call stacks in no particular order:
Call stack:
    0) onHashChange1 (async-callstack-events.html:52)
    [hashchange]
    0) doTestHashChange (async-callstack-events.html:46)
    1) testFunction (async-callstack-events.html:11)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Call stack:
    0) onHashChange2 (async-callstack-events.html:58)
    [hashchange]
    0) doTestHashChange (async-callstack-events.html:46)
    1) testFunction (async-callstack-events.html:11)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Call stack:
    0) onSelectionChange (async-callstack-events.html:39)
    [selectionchange]
    0) setSelection (async-callstack-events.html:24)
    1) doTestSelectionChange (async-callstack-events.html:31)
    2) testFunction (async-callstack-events.html:11)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>

Call stack:
    0) onVideoPlay (async-callstack-events.html:71)
    [play]
    0) doTestMediaEvents (async-callstack-events.html:65)
    1) testFunction (async-callstack-events.html:11)
    [setTimeout]
    0) scheduleTestFunction (debugger-test.js:3)
    <... skipped remaining frames ...>
